requirements.txt
flask
google-cloud-texttospeech
transformers
torch
spacy